refactor: split models and centralize utilities
