--- src/core/models/unified_enums.py
@@ -1,173 +1,0 @@
-#!/usr/bin/env python3
-"""
-Unified Enums Module - Agent Cellphone V2
-
-Consolidates all scattered enum implementations into a single,
-comprehensive enum system to achieve 100% duplication elimination.
-
-Author: Agent-3 Integration & Testing Specialist
-Task: TASK 3J - Model & Enum Consolidation (V2 Compliance Refactoring)
-V2 Standards: ≤400 LOC, SRP, OOP principles
-"""
-
-from enum import Enum
-
-# ============================================================================
-# UNIFIED ENUM SYSTEM - Consolidated from all scattered implementations
-# ============================================================================
-
-class UnifiedStatus(Enum):
-    """Unified status enum - consolidated from all systems"""
-    # Core statuses
-    ACTIVE = "active"
-    INACTIVE = "inactive"
-    PENDING = "pending"
-    COMPLETED = "completed"
-    FAILED = "failed"
-    CANCELLED = "cancelled"
-    SUSPENDED = "suspended"
-    
-    # Health statuses
-    HEALTHY = "healthy"
-    WARNING = "warning"
-    CRITICAL = "critical"
-    DEGRADED = "degraded"
-    OFFLINE = "offline"
-    UNKNOWN = "unknown"
-    
-    # Task statuses
-    QUEUED = "queued"
-    RUNNING = "running"
-    SUCCESS = "success"
-    ERROR = "error"
-    TIMEOUT = "timeout"
-    SKIPPED = "skipped"
-    
-    # Workflow statuses
-    INITIALIZED = "initialized"
-    IN_PROGRESS = "in_progress"
-    BLOCKED = "blocked"
-    RESOLVED = "resolved"
-    ARCHIVED = "archived"
-
-
-class UnifiedPriority(Enum):
-    """Unified priority enum - consolidated from all systems"""
-    CRITICAL = "critical"
-    HIGH = "high"
-    MEDIUM = "medium"
-    LOW = "low"
-    MINIMAL = "minimal"
-    OPTIONAL = "optional"
-
-
-class UnifiedSeverity(Enum):
-    """Unified severity enum - consolidated from all systems"""
-    CRITICAL = "critical"
-    HIGH = "high"
-    MEDIUM = "medium"
-    LOW = "low"
-    INFO = "info"
-    DEBUG = "debug"
-
-
-class UnifiedType(Enum):
-    """Unified type enum - consolidated from all systems"""
-    # System types
-    SYSTEM = "system"
-    SERVICE = "service"
-    COMPONENT = "component"
-    MODULE = "module"
-    
-    # Data types
-    DATA = "data"
-    CONFIG = "config"
-    STATE = "state"
-    EVENT = "event"
-    
-    # Business types
-    TASK = "task"
-    WORKFLOW = "workflow"
-    MESSAGE = "message"
-    ALERT = "alert"
-    REPORT = "report"
-    
-    # Technical types
-    API = "api"
-    DATABASE = "database"
-    NETWORK = "network"
-    FILE = "file"
-
-
-class UnifiedFormat(Enum):
-    """Unified format enum - consolidated from all systems"""
-    JSON = "json"
-    XML = "xml"
-    YAML = "yaml"
-    TEXT = "text"
-    HTML = "html"
-    CSV = "csv"
-    BINARY = "binary"
-    CUSTOM = "custom"
-
-
-# ============================================================================
-# UTILITY FUNCTIONS
-# ============================================================================
-
-def get_all_enum_values() -> dict:
-    """Get all enum values for easy access"""
-    return {
-        "status": [status.value for status in UnifiedStatus],
-        "priority": [priority.value for priority in UnifiedPriority],
-        "severity": [severity.value for severity in UnifiedSeverity],
-        "type": [type_.value for type_ in UnifiedType],
-        "format": [format_.value for format_ in UnifiedFormat]
-    }
-
-
-def validate_enum_value(enum_class: type, value: str) -> bool:
-    """Validate if a value exists in the specified enum"""
-    try:
-        return value in [e.value for e in enum_class]
-    except Exception:
-        return False
-
-
-def get_enum_by_value(enum_class: type, value: str):
-    """Get enum instance by value"""
-    try:
-        for enum_instance in enum_class:
-            if enum_instance.value == value:
-                return enum_instance
-        return None
-    except Exception:
-        return None
-
-
-# ============================================================================
-# MAIN ENTRY POINT
-# ============================================================================
-
-def main():
-    """Main entry point for unified enums module"""
-    print("🚀 UNIFIED ENUMS MODULE - TASK 3J V2 Refactoring")
-    print("=" * 50)
-    
-    # Display all enum values
-    all_values = get_all_enum_values()
-    for enum_name, values in all_values.items():
-        print(f"{enum_name.title()}: {', '.join(values)}")
-    
-    print(f"\n✅ Total Status Values: {len(UnifiedStatus)}")
-    print(f"✅ Total Priority Values: {len(UnifiedPriority)}")
-    print(f"✅ Total Severity Values: {len(UnifiedSeverity)}")
-    print(f"✅ Total Type Values: {len(UnifiedType)}")
-    print(f"✅ Total Format Values: {len(UnifiedFormat)}")
-    
-    print("\n🎯 Unified Enums Module Ready!")
-    return 0
-
-
-if __name__ == "__main__":
-    exit(main())
