--- src/core/models/unified_framework.py
@@ -1,240 +1,0 @@
-#!/usr/bin/env python3
-"""
-Unified Framework Module - Agent Cellphone V2
-
-Main entry point for the unified model framework, providing
-consolidated access to all model functionality.
-
-Author: Agent-3 Integration & Testing Specialist
-Task: TASK 3J - Model & Enum Consolidation (V2 Compliance Refactoring)
-V2 Standards: ≤400 LOC, SRP, OOP principles
-"""
-
-import logging
-from typing import Dict, List, Any, Optional
-from datetime import datetime
-
-try:
-    from src.core.models.unified_enums import UnifiedStatus, UnifiedPriority, UnifiedSeverity, UnifiedType, UnifiedFormat
-    from src.core.models.base_models import BaseModel, StatusModel, TypedModel, ConfigurableModel
-    from src.core.models.health_performance_models import HealthModel, PerformanceModel
-    from src.core.models.task_workflow_models import TaskModel, WorkflowModel
-    from src.core.models.message_models import MessageModel
-except ImportError:
-    from unified_enums import UnifiedStatus, UnifiedPriority, UnifiedSeverity, UnifiedType, UnifiedFormat
-    from base_models import BaseModel, StatusModel, TypedModel, ConfigurableModel
-    from health_performance_models import HealthModel, PerformanceModel
-    from task_workflow_models import TaskModel, WorkflowModel
-    from message_models import MessageModel
-
-logger = logging.getLogger(__name__)
-
-# ============================================================================
-# MODEL REGISTRY - Centralized model management
-# ============================================================================
-
-class ModelRegistry:
-    """Centralized registry for all model types"""
-    
-    def __init__(self):
-        self.models = {}
-        self.categories = {}
-        self._setup_default_categories()
-    
-    def _setup_default_categories(self) -> None:
-        """Setup default model categories"""
-        self.categories = {
-            "health": HealthModel,
-            "performance": PerformanceModel,
-            "task": TaskModel,
-            "workflow": WorkflowModel,
-            "message": MessageModel
-        }
-    
-    def register_model(self, model_class: type, category: str) -> bool:
-        """Register a model class in a category"""
-        try:
-            if category not in self.categories:
-                self.categories[category] = model_class
-            return True
-        except Exception as e:
-            logger.error(f"Failed to register model {model_class.__name__}: {e}")
-            return False
-    
-    def get_model_class(self, category: str) -> Optional[type]:
-        """Get model class by category"""
-        return self.categories.get(category)
-    
-    def get_registry_summary(self) -> Dict[str, Any]:
-        """Get summary of registered models"""
-        return {
-            "total_models": len(self.categories),
-            "categories": list(self.categories.keys()),
-            "model_classes": [cls.__name__ for cls in self.categories.values()]
-        }
-
-
-# ============================================================================
-# MODEL FACTORY - Factory for creating unified models
-# ============================================================================
-
-class UnifiedModelFactory:
-    """Factory for creating unified models with proper configuration"""
-    
-    def __init__(self, registry: ModelRegistry):
-        self.registry = registry
-    
-    def create_health_model(self, **kwargs) -> HealthModel:
-        """Create a health model"""
-        return HealthModel(**kwargs)
-    
-    def create_performance_model(self, **kwargs) -> PerformanceModel:
-        """Create a performance model"""
-        return PerformanceModel(**kwargs)
-    
-    def create_task_model(self, **kwargs) -> TaskModel:
-        """Create a task model"""
-        return TaskModel(**kwargs)
-    
-    def create_workflow_model(self, **kwargs) -> WorkflowModel:
-        """Create a workflow model"""
-        return WorkflowModel(**kwargs)
-    
-    def create_message_model(self, **kwargs) -> MessageModel:
-        """Create a message model"""
-        return MessageModel(**kwargs)
-    
-    def create_model(self, model_type: str, **kwargs) -> Optional[BaseModel]:
-        """Create a model by type"""
-        model_map = {
-            "health": self.create_health_model,
-            "performance": self.create_performance_model,
-            "task": self.create_task_model,
-            "workflow": self.create_workflow_model,
-            "message": self.create_message_model
-        }
-        
-        creator = model_map.get(model_type.lower())
-        if creator:
-            return creator(**kwargs)
-        return None
-
-
-# ============================================================================
-# MAIN UNIFIED MODEL FRAMEWORK - Entry point for all operations
-# ============================================================================
-
-class UnifiedModelFramework:
-    """
-    Main unified model framework - consolidates all scattered implementations
-    
-    Provides:
-    - Unified enums for all systems
-    - Base model classes with common functionality
-    - Specialized model categories
-    - Centralized model registry
-    - Factory for model creation
-    """
-    
-    def __init__(self):
-        self.registry = ModelRegistry()
-        self.factory = UnifiedModelFactory(self.registry)
-        self.logger = logging.getLogger(__name__)
-    
-    def get_registry(self) -> ModelRegistry:
-        """Get the model registry"""
-        return self.registry
-    
-    def get_factory(self) -> UnifiedModelFactory:
-        """Get the model factory"""
-        return self.factory
-    
-    def get_available_models(self) -> List[str]:
-        """Get all available model types"""
-        return ["health", "performance", "task", "workflow", "message"]
-    
-    def create_model(self, model_type: str, **kwargs) -> Optional[BaseModel]:
-        """Create a model of specified type"""
-        try:
-            return self.factory.create_model(model_type, **kwargs)
-        except Exception as e:
-            self.logger.error(f"Failed to create {model_type} model: {e}")
-            return None
-    
-    def get_model_info(self) -> Dict[str, Any]:
-        """Get comprehensive model framework information"""
-        return {
-            "framework_version": "2.0.0",
-            "available_models": self.get_available_models(),
-            "registry_summary": self.registry.get_registry_summary(),
-            "unified_enums": {
-                "status": [status.value for status in UnifiedStatus],
-                "priority": [priority.value for priority in UnifiedPriority],
-                "severity": [severity.value for severity in UnifiedSeverity],
-                "type": [type_.value for type_ in UnifiedType],
-                "format": [format_.value for format_ in UnifiedFormat]
-            }
-        }
-
-
-# ============================================================================
-# UTILITY FUNCTIONS - Helper functions for model operations
-# ============================================================================
-
-def create_unified_model(model_type: str, **kwargs) -> Optional[BaseModel]:
-    """Utility function to create unified models"""
-    framework = UnifiedModelFramework()
-    return framework.create_model(model_type, **kwargs)
-
-
-def get_unified_enums() -> Dict[str, List[str]]:
-    """Utility function to get all unified enum values"""
-    return {
-        "status": [status.value for status in UnifiedStatus],
-        "priority": [priority.value for priority in UnifiedPriority],
-        "severity": [severity.value for severity in UnifiedSeverity],
-        "type": [type_.value for type_ in UnifiedType],
-        "format": [format_.value for format_ in UnifiedFormat]
-    }
-
-
-def validate_model_data(model: BaseModel) -> bool:
-    """Utility function to validate model data"""
-    try:
-        return model.validate()
-    except Exception:
-        return False
-
-
-# ============================================================================
-# MAIN ENTRY POINT
-# ============================================================================
-
-def main():
-    """Main entry point for unified model framework"""
-    framework = UnifiedModelFramework()
-    
-    # Display framework information
-    info = framework.get_model_info()
-    print("🚀 UNIFIED MODEL FRAMEWORK - TASK 3J V2 Refactoring")
-    print("=" * 50)
-    print(f"Framework Version: {info['framework_version']}")
-    print(f"Available Models: {', '.join(info['available_models'])}")
-    print(f"Total Registered Models: {info['registry_summary']['total_models']}")
-    print(f"Categories: {', '.join(info['registry_summary']['categories'])}")
-    
-    # Create example models
-    health_model = framework.create_model("health", health_score=95.0)
-    task_model = framework.create_model("task", title="Example Task", priority=UnifiedPriority.HIGH)
-    
-    print(f"\n✅ Created Health Model: {health_model.id}")
-    print(f"✅ Created Task Model: {task_model.id}")
-    
-    print("\n🎯 Model & Enum Consolidation Framework Ready!")
-    print("✅ V2 Standards Compliance: ACHIEVED")
-    print("✅ All modules ≤400 LOC")
-    return 0
-
-
-if __name__ == "__main__":
-    exit(main())
